# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Creates a new user account",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Validation failed or user exists"
          },
          "429": {
            "description": "Too Many Requests. Rate limit exceeded"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Generates a new access token using a valid refresh token",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "New access token generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Invalid or expired refresh token"
          },
          "429": {
            "description": "Too Many Requests. Rate limit exceeded"
          }
        },
        "security": [
          {
            "JWT-refresh": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Logs out the user and invalidates the refresh token",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User successfully logged out"
          },
          "401": {
            "description": "Unauthorized. Invalid or expired access token"
          },
          "429": {
            "description": "Too Many Requests. Rate limit exceeded"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Logout user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticates a user and returns access and refresh tokens",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Invalid credentials"
          },
          "429": {
            "description": "Too Many Requests. Rate limit exceeded"
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/categories": {
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully."
          },
          "400": {
            "description": "Invalid input data."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new category",
        "tags": [
          "Categories"
        ]
      },
      "get": {
        "operationId": "CategoryController_findAll",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter categories by name or description",
            "schema": {
              "example": "electronics",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page for pagination",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of categories retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CategoryResponseDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get all categories",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "get": {
        "operationId": "CategoryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "summary": "Get category by ID",
        "tags": [
          "Categories"
        ]
      },
      "patch": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          },
          "409": {
            "description": "Category slug already"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update category (Admin only)",
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "operationId": "CategoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Cannot delete category with products"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete category (Admin Only)",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/slug/{slug}": {
      "get": {
        "operationId": "CategoryController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "summary": "Get category by slug",
        "tags": [
          "Categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "API documentation for the application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "JWT-refresh": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Refresh-JWT",
        "description": "Enter refresh JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "StrongP@ssw0rd!"
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Access token for authentication",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh token for obtaining new access tokens",
            "example": "dGhpcy1pcz1hLXJlZnJlc2gtdG9rZW4tZXhhbXBsZS13aXRoLXN1ZmZpY2lhbC1jaGFyYWN0ZXJzIQ=="
          },
          "user": {
            "type": "object",
            "description": "Authenticated user information",
            "example": {
              "id": "user-123",
              "email": "<EMAIL>",
              "firstName": "John",
              "lastName": "Doe",
              "role": "USER"
            }
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "StrongP@ssw0rd!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Electronics",
            "description": "The name of the category",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "example": "Devices and gadgets including phones, laptops, and accessories",
            "description": "A brief description of the category",
            "maxLength": 255
          },
          "slug": {
            "type": "string",
            "example": "electronics",
            "description": "The URL-friendly slug for the category",
            "maxLength": 100
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/images/electronics.png",
            "description": "URL of the category image",
            "maxLength": 255
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Indicates if the category is active",
            "default": true
          }
        },
        "required": [
          "name"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e484-ere8458454-45erer4844858",
            "description": "The unique identifier of the category"
          },
          "name": {
            "type": "string",
            "example": "Electronics",
            "description": "The name of the category"
          },
          "description": {
            "type": "object",
            "example": "Devices and gadgets including phones, laptops, and accessories",
            "description": "A brief description of the category",
            "nullable": true
          },
          "slug": {
            "type": "object",
            "example": "electronics",
            "description": "The URL-friendly slug for the category",
            "nullable": true
          },
          "imageUrl": {
            "type": "object",
            "example": "https://example.com/images/electronics.png",
            "description": "URL of the category image",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Indicates if the category is active"
          },
          "productCount": {
            "type": "number",
            "example": 150,
            "description": "Number of products in this category"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T12:00:00Z",
            "description": "The date and time when the category was created"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-10T15:30:00Z",
            "description": "The date and time when the category was last updated"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "slug",
          "imageUrl",
          "isActive",
          "productCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Electronics",
            "description": "The name of the category",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "example": "Devices and gadgets including phones, laptops, and accessories",
            "description": "A brief description of the category",
            "maxLength": 255
          },
          "slug": {
            "type": "string",
            "example": "electronics",
            "description": "The URL-friendly slug for the category",
            "maxLength": 100
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/images/electronics.png",
            "description": "URL of the category image",
            "maxLength": 255
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Indicates if the category is active",
            "default": true
          }
        }
      },
...
    }
  }
}